requirements.txt
gradio==6.0.2
google-genai==1.52.0
pypdf==6.4.0
python-dotenv==1.2.1
python-docx>=0.8.11